add HRL Laboratories
job announce:

https://us59.dayforcehcm.com/CandidatePortal/en-US/hrl/Posting/View/713

Located in Malibu, California, HRL has been on the leading edge of technology, conducting pioneering research and advancing the state of the art.

HRL is a world-leader in silicon quantum dot technologies. At HRL, we engage in the development of a full quantum device and computation stack, including an optimizing compiler for a quantum programming language.

As a Compiler Engineer within the Quantum Programming group, you will be a core contributor to this stack, including both maintenance and significant new feature development. The compiler is primarily written in Common Lisp, and interfaces with a lot of software, both as a library and as a stand-alone application. As such, in this role, you will have opportunities to work on both quantum compilers as well as open-source Lisp compilers, especially in the area of the optimization and deployment of Lisp programs.

REQUIREMENTS

    Bachelors in mathematics or computer science, or equivalent knowledge gained through industry experience

    Outstanding programming skills, including full cycle software development with git

    Expertise in Common Lisp

    Superb understanding of operating system fundamentals, the C ABI, programming language runtimes, and programming language implementation

    Excellent skills in profiling and debugging

    Good written and verbal communication; punctual; strongly self-motivated

    U.S. Citizen

PREFERRED QUALIFICATIONS

    Experience doing academic research in computer science or mathematics

    Demonstrated skills working on an industrial-strength compiler

    Knowledge of Steel Bank Common Lisp

    Familiarity with static and strong functional typing disciplines.

    Knowledge of the fundamentals of quantum computing (not necessary, but a bonus)

    Familiarity with scientific and numerical computing