Update Google/ITA section
The Google/ITA section was badly out-of-date. It would now be very misleading to say the project is mostly in Common Lisp. Not because any CL code has been converted to something else, but because the ecosystem became much larger.

Also Google may or may not be hiring at any given moment.  They're also firing. So I removed that claim.